[CL 11203920 by Marc Audy in Dev-Framework branch]

+++ Engine/Source/Developer/Windows/LiveCodingServer/Private/External/LC_NameMangling.cpp
@@ -11,19 +11,19 @@ typedef void (*free_func_t)(void*);
 extern "C" char* __cdecl __unDName(char* buffer, const char* mangled, int buflen, malloc_func_t memget, free_func_t memfree, unsigned short flags);
 
 
-std::string nameMangling::UndecorateSymbol(const char* symbolName, uint32_t flags)
+std::string nameMangling::UndecorateSymbol(const char* symbolName, uint16_t flags)
 {
 	char buffer[65536];
-	__unDName(buffer, symbolName, 65536, malloc, free, static_cast<uint16_t>(flags));
+	__unDName(buffer, symbolName, 65536, malloc, free, flags);
 
 	return std::string(buffer);
 }
 
 
-std::wstring nameMangling::UndecorateSymbolWide(const char* symbolName, uint32_t flags)
+std::wstring nameMangling::UndecorateSymbolWide(const char* symbolName, uint16_t flags)
 {
 	char buffer[65536];
-	__unDName(buffer, symbolName, 65536, malloc, free, static_cast<uint16_t>(flags));
+	__unDName(buffer, symbolName, 65536, malloc, free, flags);
 
 	return string::ToWideString(buffer);
 }
